Removed escape character recognition for strings
This was removed, because at least Coin inventor uses a strange
handling for escape characters. Although double-quotes can be
escaped by backslash, backslash characters are not escaped.
This leads to strange behavior if the string ends with a backslash,
e.g. "myString\" cannot be parsed correctly. This use case
is somehow relevant, when a string fields stores a path on Windows.